Moved old box esitmator to /api/ folder, added new box estimator using SDK and depthai-viewer
