+++ gen2-box_measurement/api/example_pcls/calibration/box.ply
