+++ gen2-box_measurement/api/example_pcls/example_2.ply
