+++ gen2-box_measurement/api/example_pcls/example_3.ply
