+++ gen2-box_measurement/api/main.py
@@ -1,0 +1,163 @@
+#!/usr/bin/env python3
+
+import time
+import cv2
+import depthai as dai
+import open3d as o3d
+import argparse
+from box_estimator import BoxEstimator
+from projector_3d import PointCloudFromRGBD
+
+
+
+parser = argparse.ArgumentParser()
+parser.add_argument('-maxd', '--max_dist', type=float, help="Maximum distance between camera and object in space in meters",
+                    default=1.5)
+parser.add_argument('-mins', '--min_box_size', type=float, help="Minimum box size in cubic meters",
+                    default=0.003)
+
+args = parser.parse_args()
+
+COLOR = True
+
+lrcheck  = True   # Better handling for occlusions
+extended = False  # Closer-in minimum depth, disparity range is doubled
+subpixel = True   # Better accuracy for longer distance, fractional disparity 32-levels
+# Options: MEDIAN_OFF, KERNEL_3x3, KERNEL_5x5, KERNEL_7x7
+median   = dai.StereoDepthProperties.MedianFilter.KERNEL_7x7
+
+print("StereoDepth config options:")
+print("    Left-Right check:  ", lrcheck)
+print("    Extended disparity:", extended)
+print("    Subpixel:          ", subpixel)
+print("    Median filtering:  ", median)
+
+pipeline = dai.Pipeline()
+
+monoLeft = pipeline.create(dai.node.MonoCamera)
+monoLeft.setResolution(dai.MonoCameraProperties.SensorResolution.THE_800_P)
+monoLeft.setBoardSocket(dai.CameraBoardSocket.LEFT)
+
+monoRight = pipeline.create(dai.node.MonoCamera)
+monoRight.setResolution(dai.MonoCameraProperties.SensorResolution.THE_800_P)
+monoRight.setBoardSocket(dai.CameraBoardSocket.RIGHT)
+
+stereo = pipeline.createStereoDepth()
+stereo.setDefaultProfilePreset(dai.node.StereoDepth.PresetMode.HIGH_ACCURACY)
+stereo.initialConfig.setMedianFilter(median)
+stereo.setLeftRightCheck(lrcheck)
+stereo.setExtendedDisparity(extended)
+stereo.setSubpixel(subpixel)
+monoLeft.out.link(stereo.left)
+monoRight.out.link(stereo.right)
+
+config = stereo.initialConfig.get()
+config.postProcessing.speckleFilter.enable = False
+config.postProcessing.speckleFilter.speckleRange = 50
+config.postProcessing.temporalFilter.enable = True
+config.postProcessing.spatialFilter.enable = True
+config.postProcessing.spatialFilter.holeFillingRadius = 2
+config.postProcessing.spatialFilter.numIterations = 1
+config.postProcessing.thresholdFilter.minRange = 400
+config.postProcessing.thresholdFilter.maxRange = 15000
+config.postProcessing.decimationFilter.decimationFactor = 1
+stereo.initialConfig.set(config)
+
+xout_depth = pipeline.createXLinkOut()
+xout_depth.setStreamName('depth')
+stereo.depth.link(xout_depth.input)
+
+# xout_disparity = pipeline.createXLinkOut()
+# xout_disparity.setStreamName('disparity')
+# stereo.disparity.link(xout_disparity.input)
+
+xout_colorize = pipeline.createXLinkOut()
+xout_colorize.setStreamName('colorize')
+if COLOR:
+    camRgb = pipeline.create(dai.node.ColorCamera)
+    camRgb.setResolution(dai.ColorCameraProperties.SensorResolution.THE_1080_P)
+    camRgb.setIspScale(1, 3)
+    camRgb.setColorOrder(dai.ColorCameraProperties.ColorOrder.RGB)
+    camRgb.initialControl.setManualFocus(130)
+    stereo.setDepthAlign(dai.CameraBoardSocket.RGB)
+    camRgb.isp.link(xout_colorize.input)
+else:
+    stereo.rectifiedRight.link(xout_colorize.input)
+
+
+class HostSync:
+    def __init__(self):
+        self.arrays = {}
+    def add_msg(self, name, msg):
+        if not name in self.arrays:
+            self.arrays[name] = []
+        # Add msg to array
+        self.arrays[name].append({'msg': msg, 'seq': msg.getSequenceNum()})
+
+        synced = {}
+        for name, arr in self.arrays.items():
+            for i, obj in enumerate(arr):
+                if msg.getSequenceNum() == obj['seq']:
+                    synced[name] = obj['msg']
+                    break
+        # If there are 3 (all) synced msgs, remove all old msgs
+        # and return synced msgs
+        if len(synced) == 2: # color, depth, nn
+            # Remove old msgs
+            for name, arr in self.arrays.items():
+                for i, obj in enumerate(arr):
+                    if obj['seq'] < msg.getSequenceNum():
+                        arr.remove(obj)
+                    else: break
+            return synced
+        return False
+
+with dai.Device(pipeline) as device:
+
+    device.setIrLaserDotProjectorBrightness(1200)
+    qs = []
+    qs.append(device.getOutputQueue("depth", 1))
+    qs.append(device.getOutputQueue("colorize", 1))
+
+    calibData = device.readCalibration()
+    if COLOR:
+        w, h = camRgb.getIspSize()
+        intrinsics = calibData.getCameraIntrinsics(dai.CameraBoardSocket.RGB, dai.Size2f(w, h))
+    else:
+        w, h = monoRight.getResolutionSize()
+        intrinsics = calibData.getCameraIntrinsics(dai.CameraBoardSocket.RIGHT, dai.Size2f(w, h))
+
+    # Sleep for 5 seconds, for the camera to settle
+    time.sleep(5)
+
+    pcl_converter = PointCloudFromRGBD(intrinsics, w, h)
+    sync = HostSync()
+    box_estimator = BoxEstimator(args.max_dist)
+
+    i = 0
+    t = time.time()
+
+    while True:
+        for q in qs:
+            new_msg = q.tryGet()
+            if new_msg is not None:
+                msgs = sync.add_msg(q.getName(), new_msg)
+                if msgs:
+                    depth = msgs["depth"].getFrame()
+                    color = msgs["colorize"].getCvFrame()
+
+                    cv2.imshow("color", color)
+                    rgb = cv2.cvtColor(color, cv2.COLOR_BGR2RGB)
+                    pointcloud = pcl_converter.rgbd_to_projection(depth, rgb)
+                    t_new = time.time()
+                    dt =  t_new - t
+                    fps = 1 / dt
+                    t = t_new
+                    l, w, h = box_estimator.process_pcl(pointcloud)
+                    if(l * w * h  > args.min_box_size):
+                        box_estimator.vizualise_box()
+                        img = box_estimator.vizualise_box_2d(intrinsics, color)
+                        cv2.imshow("2d projection", img)
+                        print(f"Length: {l:.2f}, Width: {w:.2f}, Height:{h:.2f}")
+        if cv2.waitKey(1) == ord('q'):
+            break
